Completed Week2 Task E
https://syllabus.codeyourfuture.io/react/week-2/lesson

We opened the pokedex React application and opened the CaughtPokemon.js file.
Changed the useState to be initialised to an empty array ([])
There was now be a bug in y=our app! We don't see how many Pokemon we have caught.
We changed the JSX to instead render caught.length. This fixed the bug.
Then we showed the names of the Pokemon we have caught. Rendered a <ul> element within the component. Then used the map method to loop through each item in the caught array and rendered it in an <li> element.
Changed the catchPokemon function to add a new Pokemon (it doesn't matter which one) onto the caught array. (Hint: we used the concat method.)
(STRETCH GOAL) Generated a random Pokemon each time we click the button